+++ src/Components/Modal.css
@@ -20,6 +20,7 @@ position: absolute;
   width: 100%;
 max-width: 700px ;
 
+
 }
