+++ src/Components/form.css
@@ -9,7 +9,7 @@
   padding: 2rem;
   border-radius: 0.5rem;
   background-color: #cbd3d9;
-
+overflow-y: auto;
 
    border-radius: 10px;
   box-shadow: 0px 0px 18px 0px rgba(0, 0, 0, 0.75);
