--- src/Layout_button/Button.css
@@ -1,22 +1,0 @@
-#button_to_nav{
-  border: none;
- 
-  font-size: 22px;
-
-  font-weight: bold;
-  cursor: pointer;
-  border-radius: 10px;
-  outline: none;
-  margin-left: 20px;
-  position: relative;
-  margin-top: 40px;
-  left: 45%;
-}
-
-body {
-   
-scroll-behavior: auto;
-  
-background-color: rgb(156, 190, 197);
-
-}
